+++ cli.js
@@ -2,13 +2,13 @@
 // usage: cli.js --organization "Secretaría de Cultura"
 
 const argv = require('minimist')(process.argv.slice(2))
-const contracts = require('./contracts')
+const scraper = require('./scraper')
 
 const organization = argv.organization
 
 ;(async () => {
   try {
-    await contracts.getContract(organization)
+    await scraper.getContract(organization)
     console.log('Terminamos el scraping')
   } catch (e) {
     console.log('Algo falló')
